Add WithPolicyGeneratorMethod
During usage of the original package it was found that opting to
include, for example, uppercase characters would not guarantee
inclusion of uppercase in the resulting string.  This was not the
stated intent of the initial package.  This would cause issues where
the generated password would fail against a stated password policy.

This change adds a wrapper around Generator to allow a simplistic policy
to be passed in to the generator which will only return when a password
that aligns with the policy is generated.

Signed-off-by: Richard Gee <richard@technologee.co.uk>